# optimize IPRange.Prefix[es], common prefix length, etc