@bradfitz 2020-12-30
@mdlayher, I made some slight tweaks/simplifications. Want to check again?